Editing of multiple summaries in Namespaces and Types, fixes
The Namespace view and the Type view now render their children on the
main page, allowing users to quickly edit summaries.  On namespaces,
users can quickly edit the summaries of the types contained in the
namespace, on types, users can quickly edit the summaries of the
members.

There is now a validation system that will render a Parsing error just
next to the caption, so if there is a parsing error, the user will
immediately see in which element this takes place (important for these
long namespace/type renderings, as it would be hard to pinpoint the
error otherwise).

The validate/save framework now shows an example of how you can
validate multiple nodes, track all the changes in a dirtyNodes hash,
and then iterate through those on save.