Image support, restructuring of first responder.
This now supports rendering images from the _images directory, and
adding images via a menu.

The code has been restructured, so that we no longer use the
AppDelegate as a dispatching delegate to the current window, but
instead, let the MainWindowController handle those events (prevents a
bunch of tests for mainWindowController instance, and the tracking of
this variable).

That said, the tracking is still there, untiL I remove the rest of the
dependencies.